packaging: make the installed wheel importable and runnable
`pip install openant` produced a distribution that could not start. Proven before
the fix by building a wheel, installing it into an empty venv, and running from
outside the checkout:

 ModuleNotFoundError: No module named 'experiment' (import core.analyzer)

Two defects, one cause — shipped code depending on files that are not shipped.

1. `core/analyzer.py` imported analyze_unit, parse_response and _normalize_result
 from `experiment.py`, a research harness that is a loose top-level module, not
 one of the seven packages in pyproject. Production depended on research code,
 and the dependency ran the wrong way round.

 Moved those three functions (~190 lines) into `core/analysis_core.py`. They
 were clean to extract: AST analysis showed they depend only on each other,
 stdlib, and shipped packages (utilities/, prompts/). `experiment.py` now
 imports them FROM core, inverting the direction so research depends on product.

2. `core/reporter.py` subprocessed `generate_report.py` and `export_csv.py` by
 source-tree-relative path with cwd=_CORE_ROOT. Neither shipped, and
 `openant report --format csv|html` is reachable from the CLI (cli.py:670), so
 both failed on an installed wheel. Moved into the already-shipped `report/`
 package (git mv, history preserved) and invoked as `python -m report.csv_export`
 / `report.html_report`. Dropped cwd=_CORE_ROOT: depending on the working
 directory is what made them unrunnable when installed.

Deliberately did NOT do the C1 namespace rename in this commit. An external
adjudication suggested folding it in; blast-radius analysis put it at 469 import
sites across 206 of 267 test files, and — the deciding factor — the producer-parity
guard in tests/test_hostile_repo.py resolves paths via __file__ and pytest.skips
when a file is missing, so a rename without moving tests in lockstep would
SILENTLY skip all 8 parametrizations and evaporate the boundary-marker security
invariant into a green run. Namespacing a distribution that could not import was
premature; correctness first, namespace as its own PR.

WHY 2129 TESTS COULD NOT SEE THIS. The editable install is a bare .pth containing
the checkout root, so every import resolves from source regardless of what the
wheel contains. The existing wheel test asserted only that config/languages.json
appeared in the archive — a substring check that passes wherever the file lands.
A packaging defect is structurally invisible to a suite that never leaves the
checkout.

New test builds a wheel, installs into an empty venv, and imports core.analyzer,
both report generators and the language registry from a directory OUTSIDE the
repository. It fails against the pre-fix tree.

Also caught in passing: my first pass at updating test imports used a regex broad
enough to rewrite `from utilities.dynamic_tester.reporter import generate_report`
— an unrelated function that merely shares a name — producing syntax errors.
Reverted and redone as exact-line replacements; that call site is untouched.

Verified: installed wheel imports and runs from outside the checkout, 7 languages
resolve. Suite 2129 passed, 26 skipped, exit 0 (unpiped, $? checked). ruff clean.
Go build/test clean.
